backend/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Cors==4.0.0
Flask-Caching==2.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
urllib3>=2.0.0
beautifulsoup4==4.12.2
certifi>=2023.7.22
pytest==7.4.3
black==23.11.0
flake8==6.1.0
Werkzeug==2.2.3
Jinja2==3.0.1
click==8.0.1
itsdangerous==2.0.1
MarkupSafe==2.0.1 